[web] Rename component
