+++ web/src/components/l10n/index.js
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) [2022] SUSE LLC
+ * Copyright (c) [2022-2023] SUSE LLC
  *
  * All Rights Reserved.
  *
@@ -19,8 +19,8 @@
  * find current contact information at www.suse.com.
  */
 
+export { default as InstallerLocaleSwitcher } from "./InstallerLocaleSwitcher";
 export { default as KeymapSelector } from "./KeymapSelector";
 export { default as L10nPage } from "./L10nPage";
-export { default as LanguageSwitcher } from "./LanguageSwitcher";
 export { default as LocaleSelector } from "./LocaleSelector";
 export { default as TimezoneSelector } from "./TimezoneSelector";
